CHG: Remove circular imports in /specest
- new file `const_def.py` holds all definitions needed
  in **both** the parent metha and the backend functions

Changes to be committed:
	new file:   const_def.py
	modified:   freqanalysis.py
	modified:   mtmconvol.py
	modified:   mtmfft.py
	modified:   superlet.py
	modified:   wavelet.py
